Add playGame function which manages the game and make necessary calls , also fixed few bugs
We created 'playGame' function which manages the entire
game and decide the winner after 5 rounds als0 asks
if user wants to play again.
We fixed a bug , earlier in playRound function
we printed winner without checking if there is a winner
or not , we fixed it by using if statement.